requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
plotly>=5.17.0
pyyaml>=6.0
opcua>=0.98.12
python-dotenv>=1.0.0
cryptography>=41.0.0
pyjwt>=2.8.0
python-dateutil>=2.8.2
pytz>=2023.3
watchdog>=3.0.0
scipy>=1.11.0
seaborn>=0.12.0
matplotlib>=3.7.0
openpyxl>=3.1.0